requirements.txt
# Requirements for SerialCLI
# Install with: pip install -r requirements.txt

pyserial>=3.5
